Updated Colour to work with Hex Literals

+++ Code/Framework/Core/Core/Colour.h
@@ -7,7 +7,7 @@ class Colour
 public:
 	constexpr Colour() noexcept 
 		: r(), g(), b(), a() { }
-	constexpr Colour(const uint32 rgb) noexcept
+	constexpr Colour(const int32 rgb) noexcept
 		: r(((rgb >> 16) & 0xFF) / 255.f)
 		, g(((rgb >> 8) & 0xFF) / 255.f)
 		, b(((rgb >> 0) & 0xFF) / 255.f)
@@ -44,12 +44,12 @@ public:
 	float r, g, b, a;
 };
 
-inline constexpr Colour Colour::Black((uint32)0x000000);
-inline constexpr Colour Colour::Blue((uint32)0x0000FF);
-inline constexpr Colour Colour::Cyan((uint32)0x00FFFF);
-inline constexpr Colour Colour::Gray((uint32)0x808080);
-inline constexpr Colour Colour::Green((uint32)0x00FF00);
-inline constexpr Colour Colour::Magenta((uint32)0xFF00FF);
-inline constexpr Colour Colour::Red((uint32)0xFF0000);
-inline constexpr Colour Colour::White((uint32)0xFFFFFF);
-inline constexpr Colour Colour::Yellow((uint32)0xFFFF00);
+inline constexpr Colour Colour::Black(0x000000);
+inline constexpr Colour Colour::Blue(0x0000FF);
+inline constexpr Colour Colour::Cyan(0x00FFFF);
+inline constexpr Colour Colour::Gray(0x808080);
+inline constexpr Colour Colour::Green(0x00FF00);
+inline constexpr Colour Colour::Magenta(0xFF00FF);
+inline constexpr Colour Colour::Red(0xFF0000);
+inline constexpr Colour Colour::White(0xFFFFFF);
+inline constexpr Colour Colour::Yellow(0xFFFF00);
